retry for compile

+++ compile.sh
@@ -2,8 +2,15 @@ VERSION=$(grep 'version := ' build.sbt | awk -F'"' '{print $2}')
 NAME=$(grep 'name := ' build.sbt | awk -F'"' '{print $2}')
 
 rm -rf dist
-sbt clean compile dist
-unzip target/universal/$NAME-$VERSION.zip
-mv $NAME-$VERSION dist
+
+while true; do
+	sbt clean compile dist
+
+	if [[ -e target/universal/$NAME-$VERSION.zip ]]; then
+		unzip target/universal/$NAME-$VERSION.zip > /dev/null
+		mv $NAME-$VERSION dist
+		break
+	fi
+done
 
 # sbt clean compile stage
